+++ tools/db_bench_tool.cc
@@ -27,8 +27,6 @@
 #ifdef __FreeBSD__
 #include <sys/sysctl.h>
 #endif
-#include <sys/resource.h>
-
 #include <atomic>
 #include <cinttypes>
 #include <condition_variable>
@@ -2021,40 +2019,6 @@ static void AppendWithSpace(std::string* str, Slice msg) {
   }
   str->append(msg.data(), msg.size());
 }
-
-class RequestRateLimiter {
- private:
-  // Use AutoRefillBudget to track requests per second
-  std::unique_ptr<AutoRefillBudget<size_t>> request_budget_;
-
- public:
-  // Constructor: requests_per_second is the maximum allowed requests per second
-  explicit RequestRateLimiter(size_t requests_per_second) {
-    // Refill period of 1 second (1,000,000 microseconds)
-    // Refill amount equals the requests per second limit
-    request_budget_ = std::make_unique<AutoRefillBudget<size_t>>(
-        requests_per_second, 1000000 /* 1 second in microseconds */);
-  }
-
-  // Try to process a request - returns true if allowed, false if rate limited
-  bool TryProcessRequest(size_t request_cost = 1) {
-    return request_budget_->TryConsume(request_cost);
-  }
-
-  // Get current available request budget
-  size_t GetAvailableRequests() {
-    return request_budget_->GetAvailableBudget();
-  }
-
-  // Update rate limit parameters
-  void UpdateRateLimit(size_t new_requests_per_second) {
-    request_budget_->SetRefillParameters(new_requests_per_second, 1000000);
-  }
-
-  // Reset the budget (useful for testing or manual resets)
-  void Reset() { request_budget_->Reset(); }
-};
-
 struct DBWithColumnFamilies {
   std::vector<ColumnFamilyHandle*> cfh;
   DB* db;
@@ -2265,7 +2229,6 @@ class Stats {
   uint64_t last_report_finish_;
   CPUIOUtilizationTracker usage_tracker_;
   AtomicRateTracker<size_t> req_drain_rate_;
-  ProcSysCPUUtilizationTracker proc_cpu_usage_;
   std::unordered_map<OperationType, std::shared_ptr<HistogramImpl>,
                      std::hash<unsigned char>>
       hist_;
@@ -3807,17 +3770,14 @@ class Benchmark {
       } else if (name == "fillseekseq") {
         method = &Benchmark::WriteSeqSeekSeq;
       } else if (name == "compact") {
-        fprintf(stdout, "Running compact\n");
         method = &Benchmark::Compact;
       } else if (name == "compactall") {
-        fprintf(stdout, "Detected compactall\n");
         CompactAll();
       } else if (name == "compact0") {
         CompactLevel(0);
       } else if (name == "compact1") {
         CompactLevel(1);
       } else if (name == "waitforcompaction") {
-        fprintf(stdout, "Detected waitforcompaction\n");
         WaitForCompaction();
       } else if (name == "flush") {
         Flush();
@@ -5469,7 +5429,6 @@ class Benchmark {
       batch.Clear();
       int64_t batch_bytes = 0;
       int64_t req_count = 0;
-      // for (int64_t j = 0; j < entries_per_batch_; j++) {
       for (; req_count < entries_per_batch_; req_count++) {
         int64_t rand_num = 0;
         if ((write_mode == UNIQUE_RANDOM) && (p > 0.0)) {
@@ -5588,7 +5547,6 @@ class Benchmark {
           if (req_rate_limit_->GetAvailableRequests() > 0) {
             req_rate_limit_->TryProcessRequest();
           } else {
-            // fprintf(stdout, "Request exceeded\n");
             break;
           }
         }
@@ -5699,8 +5657,6 @@ class Benchmark {
       }
       thread->stats.FinishedOps(db_with_cfh, db_with_cfh->db, req_count,
                                 kWrite);
-      // thread->stats.FinishedOps(db_with_cfh, db_with_cfh->db,
-      // entries_per_batch_, kWrite);
       if (FLAGS_sine_write_rate) {
         uint64_t now = FLAGS_env->NowMicros();
